+++ AutoSprint/AutoSprintPlugin.cs
@@ -1,9 +1,15 @@
 using System.IO;
+using System.Security;
+using System.Security.Permissions;
 using AutoSprint.Core;
 using BepInEx;
 using BepInEx.Bootstrap;
 
+[module: UnverifiableCode]
 [assembly: HG.Reflection.SearchableAttribute.OptIn]
+#pragma warning disable CS0618 // Type or member is obsolete
+[assembly: SecurityPermission(SecurityAction.RequestMinimum, SkipVerification = true)]
+#pragma warning restore CS0618 // Type or member is obsolete
 
 namespace AutoSprint
 {
@@ -14,7 +20,7 @@ namespace AutoSprint
         public const string PluginGUID = $"com.{PluginAuthor}.{PluginName}";
         public const string PluginAuthor = "score";
         public const string PluginName = "AutoSprint";
-        public const string PluginVersion = "1.4.4";
+        public const string PluginVersion = "1.4.7";
 
         internal static bool RooInstalled => Chainloader.PluginInfos.ContainsKey("com.rune580.riskofoptions");